Move Simulator class to a separate file
